requirements.txt
aiogithubapi==22.12.2
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
frozenlist==1.3.3
github==1.2.7
greenlet==2.0.1
idna==3.4
multidict==6.0.2
psycopg2==2.9.5
python-dotenv==0.21.0
requests==2.28.2
SQLAlchemy==1.4.44
typing-extensions==4.5.0
urllib3==1.26.14
yarl==1.8.1
